requirements.txt
djangorestframework-simplejwt
djangorestframework
django
phonenumber_field
phonenumbers
django-money
pillow